+++ sdk/device_update/src/client.rs
@@ -3,7 +3,7 @@ use azure_core::{
     auth::{TokenCredential, TokenResponse},
     Error as CoreError, HttpError,
 };
-use azure_identity::token_credentials::AutoRefreshingTokenCredential;
+use azure_identity::AutoRefreshingTokenCredential;
 use const_format::formatcp;
 use serde::de::DeserializeOwned;
 use std::sync::Arc;
@@ -18,7 +18,7 @@ pub(crate) const API_VERSION_PARAM: &str = formatcp!("api-version={}", API_VERSI
 ///
 /// ```no_run
 /// use azure_device_update::DeviceUpdateClient;
-/// use azure_identity::token_credentials::DefaultAzureCredential;
+/// use azure_identity::DefaultAzureCredential;
 /// let creds = std::sync::Arc::new(DefaultAzureCredential::default());
 /// let client = DeviceUpdateClient::new("contoso.api.adu.microsoft.com", creds).unwrap();
 /// ```
@@ -37,7 +37,7 @@ impl DeviceUpdateClient {
     ///
     /// ```no_run
     /// use azure_device_update::DeviceUpdateClient;
-    /// use azure_identity::token_credentials::DefaultAzureCredential;
+    /// use azure_identity::DefaultAzureCredential;
     /// let creds = std::sync::Arc::new(DefaultAzureCredential::default());
     /// let client = DeviceUpdateClient::new("contoso.api.adu.microsoft.com", creds).unwrap();
     /// ```