+++ sdk/identity/src/token_credentials/mod.rs
@@ -6,7 +6,7 @@
 //! * Managed identity
 //! * Client secret
 mod auto_refreshing_credentials;
-mod cli_credentials;
+mod azure_cli_credentials;
 #[cfg(feature = "client_certificate")]
 mod client_certificate_credentials;
 mod client_secret_credentials;
@@ -15,7 +15,7 @@ mod environment_credentials;
 mod imds_managed_identity_credentials;
 
 pub use auto_refreshing_credentials::*;
-pub use cli_credentials::*;
+pub use azure_cli_credentials::*;
 #[cfg(feature = "client_certificate")]
 pub use client_certificate_credentials::*;
 pub use client_secret_credentials::*;